Modularize conditional plugin directory structure
