+++ src_js/conditional_plugins/utils/print_enabling_url_to_console.ts
@@ -1,0 +1,8 @@
+export function printEnablingURLToConsole(
+  pluginId: string,
+  message: string,
+): void {
+  const enabled_url = new URL(window.location.href);
+  enabled_url.searchParams.set(`enable_${pluginId}`, '1');
+  console.info(`${message}\n`, enabled_url.toString());
+}
